Refactor: [BADA-84] 경로 수정 #35
