+++ src/shared/components/ui/FlatTab/FlatTab.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { FlatTab } from '@ui/FlatTab/FlatTab';
+import { FlatTab } from '@ui/FlatTab';
 
 const sampleItems = [
   {
