+++ src/shared/components/ui/SectionDivider/SectionDivider.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { SectionDivider } from '@ui/SectionDivider/SectionDivider';
+import { SectionDivider } from '@ui/SectionDivider';
 
 const meta: Meta<typeof SectionDivider> = {
   title: 'Components/SectionDivider',
